requirements.txt
python-dotenv
aiogram
yt-dlp
yandex_cloud_ml_sdk
youtube_transcript_api